Move pagination adapters into data_table/ and rename to avoid conflicts
Rename DataTablePaginationAdapters::{Pagy,Kaminari,Manual} to
DataTable{Pagy,Kaminari,Manual}Adapter under RubyUI namespace.

Fixes class name collision when host app includes RubyUI module
(e.g. Pagy clashing with the pagy gem). Also fixes generator
only copying data_table/ folder, missing the separate adapters dir.
