+++ test/ruby_ui/data_table_pagy_adapter_test.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require "ruby_ui/data_table_pagination_adapters/pagy"
+require "ruby_ui/data_table/data_table_pagy_adapter"
 
-class RubyUI::DataTablePaginationAdapters::PagyTest < ComponentTest
+class RubyUI::DataTablePagyAdapterTest < ComponentTest
   PagyDouble = Data.define(:page, :pages, :count, :items)
 
   def test_reads_page_pages_count_items
     pagy = PagyDouble.new(page: 2, pages: 5, count: 47, items: 10)
-    adapter = RubyUI::DataTablePaginationAdapters::Pagy.new(pagy)
+    adapter = RubyUI::DataTablePagyAdapter.new(pagy)
     assert_equal 2, adapter.current_page
     assert_equal 5, adapter.total_pages
     assert_equal 47, adapter.total_count
